refactor: move all form-related logic to new FormTypeTrait
